Name the sentinel index, and stop casting inside comparisons.
66 signed/unsigned comparison diagnostics, and the check's own suggestion --
std::cmp_less and friends -- was the right fix for none of them. What it
actually points at is a cast someone wrote to make a comparison well-defined,
which means the fix belongs where the types are decided, not at the
comparison.

Half of them were one thing: `(uint32_t)-1`, the "no such node / instance /
light / material" value, written out sixty-one times with no name. It is now
oka::kInvalidIndex. Same value, and the signed-to-unsigned conversion that
made every one of those a mixed comparison is gone with the spelling.

The loops that counted `for (int i = 0; i < (int)v.size(); ++i)` count in the
container's own type now, which also drops the (size_t)i casts inside them. A
glTF index that is signed because -1 means absent keeps its `>= 0` guard and
converts once past it, where the guard is visible. MetalAccelStructure's
`(size_t)-1` gets the same treatment as the scene's: kNoCurveBlas, declared
next to the function whose contract it is.

Two loops compared a dimension against an unsigned bound; they clamp the
bound into int once, by name. A texture edge that does not fit in an int is
past every limit a GPU has, so the clamp cannot change a decision.

Nothing here uses std::cmp_*. The check is at zero, and every site reads
better than it did rather than merely differently.

+++ src/render/metal/MetalRender.mm
@@ -1382,9 +1382,9 @@ void MetalRender::render(Buffer* output)
         mAnimTargetTimes.resize(animCount);
         mAnimChanged.resize(animCount);
         std::fill(mAnimChanged.begin(), mAnimChanged.end(), false);
-        for (int i = 0; i < (int)animCount; ++i)
+        for (size_t i = 0; i < animCount; ++i)
         {
-            mAnimTargetTimes[i] = animSettings.getAs<float>(animationTimeKey((size_t)i));
+            mAnimTargetTimes[i] = animSettings.getAs<float>(animationTimeKey(i));
             const float delta = std::abs(animations[i].current - mAnimTargetTimes[i]);
             if (delta > EPSILON)
             {
@@ -1414,7 +1414,7 @@ void MetalRender::render(Buffer* output)
         // predecessor to reproject from. Playback advances a sixtieth of a second
         // at a time, so a fraction of the clip length separates the two cases by a
         // wide margin. This was already being computed and then not used.
-        for (int i = 0; i < (int)animCount; ++i)
+        for (size_t i = 0; i < animCount; ++i)
         {
             const float clip = animations[i].end - animations[i].start;
             if (mAnimChanged[i] && clip > 0.0f && maxTimeDelta > 0.05f * clip)
@@ -1454,7 +1454,7 @@ void MetalRender::render(Buffer* output)
 
                 // --- Pass 1: evaluate CHANGED animations at t_open ---
                 bool pass1Skeletal = false;
-                for (int i = 0; i < (int)animations.size(); ++i)
+                for (size_t i = 0; i < animations.size(); ++i)
                 {
                     float tOpen = mAnimTargetTimes[i];
                     if (mAnimChanged[i])
@@ -1485,7 +1485,7 @@ void MetalRender::render(Buffer* output)
 
                 // --- Pass 2: evaluate CHANGED animations at t_close ---
                 bool pass2Skeletal = false;
-                for (int i = 0; i < (int)animations.size(); ++i)
+                for (size_t i = 0; i < animations.size(); ++i)
                 {
                     float tClose = mAnimTargetTimes[i];
                     if (mAnimChanged[i])
@@ -1507,14 +1507,14 @@ void MetalRender::render(Buffer* output)
                 encodeTlas = true;
 
                 // Restore target times so next-frame EPSILON check is stable
-                for (int i = 0; i < (int)animations.size(); ++i)
+                for (size_t i = 0; i < animations.size(); ++i)
                     animations[i].current = mAnimTargetTimes[i];
             }
             else
             {
                 // Motion blur disabled or no shutter: single-pass at target time
                 bool accelStructureDirty = false;
-                for (int i = 0; i < (int)animations.size(); ++i)
+                for (size_t i = 0; i < animations.size(); ++i)
                 {
                     animations[i].current = mAnimTargetTimes[i];
                     accelStructureDirty |= mScene->applyAnimation(i);
